Rename WrapperTypeContainer to WrappedTypeContainer; move class outside of CodeGen namespace
